ContactBookV2.py
Functionality: 


Add Contacts: When the user chooses the 'Add (a)' option, the program prompts the user to input a name and a phone number. These details are then written to a file named contact.txt in the format name : number. 

Search Contacts: When the user chooses the 'Search (s)' option, the program prompts the user to input a search term (either a name or a number).The program opens the contact.txt file in read mode and iterates through each line to check if the search term is present. If a match is found, the matching line is printed. If no matches are found, a message is printed to inform the user. 

Quit: The user can quit the program by choosing the 'Quit (q)' option.  

Invalid Choice Handling: If the user inputs an invalid option, the program informs the user and prompts them to try again.